Move block_annotate from iomgr to support since it's used in other core places besides iomgr
