+++ vsprojects/vcxproj/grpc_unsecure/grpc_unsecure.vcxproj
@@ -256,7 +256,6 @@
     <ClInclude Include="..\..\..\src\core\iomgr\alarm.h" />
     <ClInclude Include="..\..\..\src\core\iomgr\alarm_heap.h" />
     <ClInclude Include="..\..\..\src\core\iomgr\alarm_internal.h" />
-    <ClInclude Include="..\..\..\src\core\iomgr\block_annotate.h" />
     <ClInclude Include="..\..\..\src\core\iomgr\endpoint.h" />
     <ClInclude Include="..\..\..\src\core\iomgr\endpoint_pair.h" />
     <ClInclude Include="..\..\..\src\core\iomgr\fd_posix.h" />
